changed to settings to settingsPage
