+++ frontend/src/main-page/MainPage.tsx
@@ -4,7 +4,7 @@ import GrantPage from "./grants/GrantPage";
 import NavBar from "./navbar/NavBar";
 import RestrictedPage from "./restricted/RestrictedPage";
 import CashFlowPage from "./cash-flow/CashFlowPage";
-import Settings from "./settings/Settings";
+import Settings from "./settings/SettingsPage";
 import Footer from "../Footer";
 import UsersPage from "./users/UsersPage";
